[refactor] 사용자 정의 예외 공통 디렉토리로 옮기기
