+++ src/main/java/nextstep/subway/common/exception/NoSuchDataException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.exception;
+package nextstep.subway.common.exception;
 
 public class NoSuchDataException extends RuntimeException {
