+++ src/main/java/nextstep/subway/line/domain/Sections.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.line.exception.InvalidDataException;
+import nextstep.subway.common.exception.InvalidDataException;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.CascadeType;
